requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Medical imaging
monai>=1.3.0

# Training utilities
ranger-adabelief>=0.12.1
wandb>=0.16.0

# Vision model components
timm>=0.9.0

# Visualization
matplotlib>=3.7.0
numpy>=1.24.0

# Notebook
ipykernel>=6.0.0
ipywidgets>=8.0.0
